+++ src/ProbabilityMeasures.jl
@@ -27,6 +27,7 @@ include("core/interface.jl")
 include("univariate/continuous/normal.jl")
 include("univariate/continuous/lognormal.jl")
 include("univariate/continuous/exponential.jl")
+include("univariate/continuous/weibull.jl")
 include("univariate/continuous/uniform.jl")
 include("univariate/continuous/laplace.jl")
 include("univariate/continuous/cauchy.jl")
@@ -74,6 +75,7 @@ export mean, var, std, median, quantile, cov
 export Normal
 export LogNormal
 export Exponential
+export Weibull
 export Uniform
 export Laplace
 export Cauchy